fix(ffi): align sa_family_t to uint16_t, fix hardcoded addr_len
socket.moon defined sa_family_t as unsigned int (32-bit), conflicting
with ffi_defs.moon which correctly uses uint16_t per POSIX/Linux
x86-64 (<bits/socket.h>). The mismatch caused the 4 mac_learner_ipc
integration tests to be skipped whenever socket.lua was loaded first
in the test suite (ffi.cdef is irreversible within a LuaJIT session).

- src/lib/socket.moon: typedef uint16_t sa_family_t (was unsigned int)
- src/mac_learner_ipc.moon: use ffi.offsetof("struct sockaddr_un",
  "sun_path") instead of the hardcoded constant 2 for addr_len

Result: make test goes from 487/0/0/4 to 491/0/0/0.

+++ lua/lib/socket.lua
@@ -3,7 +3,7 @@ local log_debug
 log_debug = require("log").log_debug
 pcall(function()
   return ffi.cdef([[    typedef int socklen_t;
-    typedef unsigned int sa_family_t;
+    typedef uint16_t sa_family_t;
     typedef unsigned int ssize_t;
 
     struct sockaddr {
